DCW - Code to upgrade secrets file to new format
mainloop() now checks the format of the secrets data structure and can
upgrade to latest format if desired.

mainloop() now checks how long it's been since the last time the master
passphrase was entered and will prompt for it again if required.

Added get_yes_or_no() which loops on a prompt till a yes or no is given.

Added is_current_format() which checks for the old format.

Added upgrade_secrets_file() which loads, upgrades, and saves secrets
file.